[closes #77] u64 to usize

+++ kernel-rs/src/console.rs
@@ -71,16 +71,16 @@ const INPUT_BUF: usize = 128;
 static mut cons: Console = Console::zeroed();
 
 /// user write()s to the console go here.
-unsafe fn consolewrite(mut user_src: i32, mut src: u64, mut n: i32) -> i32 {
+unsafe fn consolewrite(mut user_src: i32, mut src: usize, mut n: i32) -> i32 {
     cons.lock.acquire();
     for i in 0..n {
         let mut c: libc::c_char = 0;
         if either_copyin(
             &mut c as *mut libc::c_char as *mut libc::c_void,
             user_src,
-            src.wrapping_add(i as u64),
-            1 as i32 as u64,
-        ) == -(1 as i32)
+            src.wrapping_add(i as usize),
+            1usize,
+        ) == -1
         {
             break;
         }
@@ -94,16 +94,16 @@ unsafe fn consolewrite(mut user_src: i32, mut src: u64, mut n: i32) -> i32 {
 /// copy (up to) a whole input line to dst.
 /// user_dist indicates whether dst is a user
 /// or kernel address.
-unsafe fn consoleread(mut user_dst: i32, mut dst: u64, mut n: i32) -> i32 {
+unsafe fn consoleread(mut user_dst: i32, mut dst: usize, mut n: i32) -> i32 {
     let mut target: u32 = n as u32;
     cons.lock.acquire();
-    while n > 0 as i32 {
+    while n > 0 {
         // wait until interrupt handler has put some
         // input into cons.buffer.
         while cons.r == cons.w {
             if (*myproc()).killed != 0 {
                 cons.lock.release();
-                return -(1 as i32);
+                return -1;
             }
             sleep(&mut cons.r as *mut u32 as *mut libc::c_void, &mut cons.lock);
         }
@@ -126,8 +126,8 @@ unsafe fn consoleread(mut user_dst: i32, mut dst: u64, mut n: i32) -> i32 {
                 user_dst,
                 dst,
                 &mut cbuf as *mut libc::c_char as *mut libc::c_void,
-                1 as i32 as u64,
-            ) == -(1 as i32)
+                1usize,
+            ) == -1
             {
                 break;
             }
@@ -159,10 +159,7 @@ pub unsafe fn consoleintr(mut cin: i32) {
         // Kill line.
         m if m == ctrl('U') => {
             while cons.e != cons.w
-                && cons.buf[cons
-                    .e
-                    .wrapping_sub(1 as i32 as u32)
-                    .wrapping_rem(INPUT_BUF as u32) as usize] as i32
+                && cons.buf[cons.e.wrapping_sub(1).wrapping_rem(INPUT_BUF as u32) as usize] as i32
                     != '\n' as i32
             {
                 cons.e = cons.e.wrapping_sub(1);
@@ -178,7 +175,7 @@ pub unsafe fn consoleintr(mut cin: i32) {
             }
         }
         _ => {
-            if cin != 0 as i32 && cons.e.wrapping_sub(cons.r) < INPUT_BUF as u32 {
+            if cin != 0 && cons.e.wrapping_sub(cons.r) < INPUT_BUF as u32 {
                 cin = if cin == '\r' as i32 { '\n' as i32 } else { cin };
 
                 // echo back to the user.
@@ -211,7 +208,7 @@ pub unsafe fn consoleinit() {
     // connect read and write system calls
     // to consoleread and consolewrite.
     let fresh2 = &mut (*devsw.as_mut_ptr().offset(CONSOLE)).read;
-    *fresh2 = Some(consoleread as unsafe fn(_: i32, _: u64, _: i32) -> i32);
+    *fresh2 = Some(consoleread as unsafe fn(_: i32, _: usize, _: i32) -> i32);
     let fresh3 = &mut (*devsw.as_mut_ptr().offset(CONSOLE)).write;
-    *fresh3 = Some(consolewrite as unsafe fn(_: i32, _: u64, _: i32) -> i32);
+    *fresh3 = Some(consolewrite as unsafe fn(_: i32, _: usize, _: i32) -> i32);
 }
